+++ sqlite3pp/sqlite3ppext.cpp
@@ -132,7 +132,7 @@ namespace sqlite3pp
 
     void context::result(char const* value, bool fstatic)
     {
-      sqlite3_result_text(ctx_, value, strlen(value), fstatic ? SQLITE_STATIC : SQLITE_TRANSIENT);
+      sqlite3_result_text(ctx_, value, static_cast<int>(strlen(value)), fstatic ? SQLITE_STATIC : SQLITE_TRANSIENT);
     }
 
     void context::result(void const* value, int n, bool fstatic)
@@ -157,7 +157,7 @@ namespace sqlite3pp
 
     void context::result_error(char const* msg)
     {
-      sqlite3_result_error(ctx_, msg, strlen(msg));
+      sqlite3_result_error(ctx_, msg, static_cast<int>(strlen(msg)));
     }
 
     void* context::aggregate_data(int size)
